% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reducedNYTS2021data}
\alias{reducedNYTS2021data}
\title{Reduced NYTS data set}
\format{
## `reducedNYTS2021data`
A data frame with 1150 rows and 24 columns:
\describe{
  \item{psu2}{Primary sampling unit}
  \item{v_stratum2}{Strata information}
  ...
}
}
\source{
<https://www.cdc.gov/tobacco/data_statistics/surveys/nyts/data/index.html>
}
\usage{
reducedNYTS2021data
}
\description{
A subset of data from the National Youth Tobacco
 Survey (NYTS) Study
}
\keyword{datasets}
