% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reducedPATHdata}
\alias{reducedPATHdata}
\title{Reduced PATH tobacco use data set}
\format{
## `reducedPATHdata`
A data frame with 132 rows and 107 columns:
\describe{
  \item{PERSONID}{Participant ID}
  \item{R01_AC1002}{Ever smoked a cigarette}
  ...
}
}
\source{
<https://www.icpsr.umich.edu/web/NAHDAP/studies/36498>
}
\usage{
reducedPATHdata
}
\description{
A subset of data from the Population Assessment of
 Tobacco and Health (PATH) Study
}
\keyword{datasets}
